feat: Implement `Watchable::has_watchers` and fix a couple of edge cases (#7)
## Description

This allows you to check if your `Watchable` is still watched by any
other values.

Also I noticed that `Direct::poll_updated` wouldn't update the (now
stored) state, so `get` was able to return outdated values when the
watchable was dropped in between. That'd be slightly weird. Also added a
test for this.

Aaaand I also noticed that `Watcher::updated` wouldn't be woken up when
the last `Watchable` was dropped. Now we `impl Drop for Watchable` and
wake up all wakers in there. This might cause wakeups when it's not the
*last* watchable that's updated, so I'm now additionally checking that
in `Direct::poll_updated`. Also wrote tests for these things.

## Change checklist

- [x] Self-review.
- [x] Documentation updates following the [style
guide](https://rust-lang.github.io/rfcs/1574-more-api-documentation-conventions.html#appendix-a-full-conventions-text),
if relevant.
- [x] Tests if relevant.