fix: Free the allocation the Weak pointer keeps alive once we notice its strong count is 0 (#38)
## Description

This allows us to avoid keeping around a potentially big allocation that
won't be reachable anymore, only because the `Weak` pointer is still
kept alive (see the weak struct's rustdoc).

Also: Change the `Drop` impl to be on `Shared` instead of `Watchable`,
so that we only notify about disconnection once when the last
`Watchable` is dropped, not every time any watchable is dropped. This
avoids some suprious wakeups.

## Breaking Changes

None